test(core): split watch sourcemap rebuild tests
